11-Flex/api/requirements.txt
langgraph>=0.0.25
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-core>=0.1.20
python-dotenv>=1.0.0
fastapi>=0.110.0
pydantic>=2.5.0
uvicorn>=0.25.0
asyncio>=3.4.3
talib-binary>=0.4.26
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0 